ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.21.0
Pillow>=10.4.0
numpy>=1.26.4
